Move Product interface to the types folder
