+++ ui/src/Products/ArchivedProduct.tsx
@@ -16,10 +16,10 @@
  */
 
 import React from 'react';
-import {Product} from './Product';
 import {useSetRecoilState} from 'recoil';
 import {ModalContentsState} from '../State/ModalContentsState';
 import ProductForm from './ProductForm';
+import {Product} from '../Types/Product';
 
 interface Props{
     product: Product;
