+++ ui/src/Products/ArchivedProductsDrawer.tsx
@@ -17,11 +17,12 @@
 
 import React, {useState} from 'react';
 import ArchivedProduct from './ArchivedProduct';
-import {isArchivedOnDate, Product} from './Product';
+import {isArchivedOnDate} from './ProductService';
 import DrawerContainer from '../ReusableComponents/DrawerContainer';
 import {useRecoilValue} from 'recoil';
-import {ViewingDateState} from '../State/ViewingDateState';
-import {ProductsState} from '../State/ProductsState';
+import {ViewingDateState} from 'State/ViewingDateState';
+import {ProductsState} from 'State/ProductsState';
+import {Product} from 'Types/Product';
 
 import '../Styles/Main.scss';
 import './ArchivedProductsDrawer.scss';
