+++ ui/src/Products/ProductClient.ts
@@ -16,11 +16,11 @@
  */
 
 import Axios, {AxiosResponse} from 'axios';
-import {Product} from './Product';
 import moment from 'moment';
 import {getToken} from '../Auth/TokenProvider';
 import MatomoEvents from '../Matomo/MatomoEvents';
 import {Space} from 'Types/Space';
+import {Product} from '../Types/Product';
 
 class ProductClient {
     private static getBaseProductsUrl(spaceUuid: string): string {
