+++ ui/src/Products/ProductForm.tsx
@@ -19,23 +19,24 @@ import React, {ChangeEvent, FormEvent, useState} from 'react';
 import {JSX} from '@babel/types';
 import moment from 'moment';
 import ProductClient from './ProductClient';
-import {emptyProduct, Product} from './Product';
+import {emptyProduct} from './ProductService';
 import ConfirmationModal, {ConfirmationModalProps} from 'Modal/ConfirmationModal/ConfirmationModal';
 import {Tag} from 'Types/Tags';
 import ProductFormLocationField from './ProductFormLocationField';
 import ProductFormStartDateField from './ProductFormStartDateField';
 import ProductFormEndDateField from './ProductFormEndDateField';
-import FormNotesTextArea from '../ModalFormComponents/FormNotesTextArea';
-import FormButton from '../ModalFormComponents/FormButton';
+import FormNotesTextArea from 'ModalFormComponents/FormNotesTextArea';
+import FormButton from 'ModalFormComponents/FormButton';
 import 'react-datepicker/dist/react-datepicker.css';
 import './ProductForm.scss';
-import {MetadataReactSelectProps} from '../ModalFormComponents/SelectWithCreateOption';
-import ProductTagClient from '../Tags/ProductTag/ProductTagClient';
-import FormTagsField from '../ReusableComponents/FormTagsField';
+import {MetadataReactSelectProps} from 'ModalFormComponents/SelectWithCreateOption';
+import ProductTagClient from 'Tags/ProductTag/ProductTagClient';
+import FormTagsField from 'ReusableComponents/FormTagsField';
 import {useRecoilValue, useSetRecoilState} from 'recoil';
-import {ViewingDateState} from '../State/ViewingDateState';
-import {ModalContentsState} from '../State/ModalContentsState';
-import {CurrentSpaceState} from '../State/CurrentSpaceState';
+import {ViewingDateState} from 'State/ViewingDateState';
+import {ModalContentsState} from 'State/ModalContentsState';
+import {CurrentSpaceState} from 'State/CurrentSpaceState';
+import {Product} from 'Types/Product';
 
 interface Props {
     editing: boolean;
