+++ ui/src/Products/ProductFormEndDateField.tsx
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021 Ford Motor Company
+ * Copyright (c) 2022 Ford Motor Company
  * All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -19,7 +19,7 @@ import DatePicker from 'react-datepicker';
 import moment from 'moment';
 import MaskedInput from 'react-text-mask';
 import React, {useState} from 'react';
-import {Product} from './Product';
+import {Product} from '../Types/Product';
 
 interface Props {
     currentProduct: Product;
