+++ ui/src/Products/ProductFormStartDateField.tsx
@@ -19,9 +19,9 @@ import DatePicker from 'react-datepicker';
 import moment from 'moment';
 import MaskedInput from 'react-text-mask';
 import React, {useState} from 'react';
-import {Product} from './Product';
 import {useRecoilValue} from 'recoil';
 import {ViewingDateState} from '../State/ViewingDateState';
+import {Product} from '../Types/Product';
 
 interface Props {
     currentProduct: Product;
