+++ ui/src/ReusableComponents/ProductCardArray.tsx
@@ -14,9 +14,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import {Product} from '../Products/Product';
 import ProductCard from '../Products/ProductCard';
 import React from 'react';
+import {Product} from '../Types/Product';
 
 interface ProductCardArrayProps {
     products: Array<Product>;
