+++ ui/src/Utils/TestData.ts
@@ -15,14 +15,14 @@
  * limitations under the License.
  */
 
-import {Person} from '../People/Person';
-import {Assignment} from '../Assignments/Assignment';
-import {Product} from '../Products/Product';
+import {Person} from 'People/Person';
+import {Assignment} from 'Assignments/Assignment';
 import {LocationTag, RoleTag, Tag} from 'Types/Tags';
 import {Space} from 'Types/Space';
-import {UserSpaceMapping} from '../Types/UserSpaceMapping';
-import {LocalStorageFilters} from '../SortingAndFiltering/FilterLibraries';
-import {Color} from '../Types/Color';
+import {UserSpaceMapping} from 'Types/UserSpaceMapping';
+import {LocalStorageFilters} from 'SortingAndFiltering/FilterLibraries';
+import {Color} from 'Types/Color';
+import {Product} from 'Types/Product';
 
 const originDateString = '2019-01-01';
